更新运行case的逻辑，

+++ my_python_code/myCase/appium_case/快快教瘦/test_case/燃脂循环2课程第四期_第九节_单台PAD.py
@@ -17,8 +17,8 @@ logging.basicConfig(level=logging.ERROR)
 logger = logging.getLogger(__name__)
 logger.info(str(os.path.basename(sys.argv[0]).split(".")[0] ))
 rult_list=[]
-def  appium_data(context,Q,course,x):
-    global rult_list, sigu
+def  appium_data(context,Q,course,x,rult_list):
+    global sigu
     try:
         system = context['platformName']
         platformVersion = context['platformVersion']
@@ -49,17 +49,17 @@ def  appium_data(context,Q,course,x):
         operate.end_courses()
         c = driver.find_elements_by_id('com.kk.coachpad:id/tv_look_report')
         assert len(c) == 9, '生成报告成功'
-        rult_list.append({'course_code': course['course_code'] + str(x), "result": 1, })
+        rult_list.put([{'course_code': course['course_code'] + str(x), "result": 1, }])
         logger.info(str({'course_code': course['course_code'] + str(x), "result": 1, }))
         return
     except Exception as e:
         sigu = True
-        rult_list.append(
-            {'course_code': course['course_code'] + str(x), "result": -1, "error_info": '\'' + str(e) + '\''})
-        Q.put(1)
-        Q.put(1)
-        Q.put(1)
-        Q.put(1)
+        rult_list.put(
+            [{'course_code': course['course_code'] + str(x), "result": -1, "error_info": '\'' + str(e) + '\''}])
+        Q.put(0)
+        Q.put(0)
+        Q.put(0)
+        Q.put(0)
 
         return
 
@@ -70,27 +70,30 @@ def test_case(context):
     phone = context['phone']
     my_db = orm_to_mysql(my_sql_link_pool())
     course=my_db.table('web_platform_course').select(name__LIKE='燃脂循环2.0课程第四期').one()
-    global rult_list,sigu
+    global sigu
     sigu=False
     Q = Queue()
+    rult_list=Queue()
     Q_pid = Queue()
   #  print(course)
     L.acquire()
-    p1 = Process(target=appium_data, args=(context,Q,course,9)) #第四个参数是上课的节数
+    p1 = Process(target=appium_data, args=(context,Q,course,9,rult_list)) #第四个参数是上课的节数
     p2 = Process(target=fictitious_bind, args=(False,32,1,Q_pid,))
     p1.start()
-    Q.get()
+    if Q.get() == 0:
+        return {"result": -1, "error_info": '\'' + str(rult_list.get()) + '\''}
     p2.start()
     p1.join()
     pid_list=Q_pid.get()
     for y in pid_list:
         os.popen('taskkill.exe /pid:' + str(y) + '   -t -f ')
     time.sleep(5)
-    Q.get()
+    if Q.get() == 0:
+        return {"result": -1, "error_info": '\'' + str(rult_list.get()) + '\''}
     L.release()
     if sigu:
-        logger.info(str({"result": -1, "error_info": '\'' + str( rult_list ) + '\''}))
-        return {"result": -1, "error_info": '\'' + str( rult_list ) + '\''}
+        logger.info(str({"result": -1, "error_info": '\'' + str( rult_list.get() ) + '\''}))
+        return {"result": -1, "error_info": '\'' + str( rult_list.get() ) + '\''}
     else:
         logger.info(str({'driverName': str(context['deviceName']), "result": 1, }))
         return {'driverName': str(context['deviceName']), "result": 1, }
